palapelify: Parallelize adjacency computation
This is the bulk of the work that is done, and is not at this point
trivial, particularly because geo::LineString and other such types have
Vec<Coordinate<T>> inner types.  This means that it is not a O(1)
operation to compute the intersection of two geometries, which means
the overall complexity is polynomial.

Of course, the "solution" to this would be to have some kind of HashSet-
based Geometry type that is suited to computing intersections quickly,
but in lieu of that, parallelizing the problem is a good first step at
speeding things up.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>