dataset: Add basic validation logic
This commit adds a basic validation function to the `CensusDatasetSchema`
struct. This function is intended to be used to validate the contents of
a dataset against the schema, and will be used in the future to validate
the contents of the dataset before attempting to load it into the
database.

This commit also adds a `CensusDataSummaryLevel` struct, which is used
to represent the summary levels in the geographic header file. This
struct is used to validate the summary levels in the geographic header
file.

This commit also adds a `CensusSummaryLevelRequirement` enum, which is
used to represent the requirements for a field to appear at a given
summary level. This is used to validate the summary levels in the
geographic header file.

This commit also adds a `CensusDataFieldSpec` enum, which is used to
represent the specification for a field in the dataset. This is used to
validate the fields in the geographic header file.

This commit also adds a `CensusDataFieldType` enum, which is used to
represent the type of data that a field in the dataset can contain. This
is used to validate the fields in the geographic header file.

This commit also adds a `CensusDataTableSpec` struct, which is used to
represent the specification for a table in the dataset. This is used to
validate the tables in the dataset.

This commit also adds a `CensusDatasetSchema` struct, which is used to
represent the schema for a dataset. This is used to validate the dataset
against the schema.

Signed-off-by: Kristofer Rye <kristofer.rye@gmail.com>